# swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Todo API",
    "description": "My todo API",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "Get index.html",
        "responses": {
          "200": {
            "description": "index.html"
          }
        }
      }
    },
    "/todos": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "newTodo": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/todos/{index}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "index",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "Todo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "1"
        },
        "text": {
          "type": "string",
          "example": "test"
        },
        "done": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "Todos": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Todo"
      }
    },
    "Text": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string",
          "example": "test"
        }
      }
    },
    "Changes": {
      "type": "object",
      "properties": {
        "changes": {
          "type": "object",
          "properties": {
            "text": {
              "type": "string",
              "example": "test"
            },
            "done": {
              "type": "boolean",
              "example": true
            }
          }
        }
      }
    }
  }
}